July18Classwork

+++ Source/Engine/Renderer/Renderer.cpp
@@ -32,7 +32,7 @@ void neu::Renderer::EndFrame()
 	SDL_RenderPresent(m_renderer);
 }
 
-void neu::Renderer::SetColor(int r, int g, int b, int a)
+void neu::Renderer::SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 {
 	SDL_SetRenderDrawColor(m_renderer, r, g, b, a);
 }
